requirements.txt
asgiref==3.3.4; python_version >= "3.6"
django-environ==0.4.5
django==3.2; python_version >= "3.6"
djangorestframework==3.12.4; python_version >= "3.5"
faker==8.1.1; python_version >= "3.6"
python-dateutil==2.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
six==1.15.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
text-unidecode==1.3; python_version >= "3.6"
